Update comment part

+++ src/components/Comment.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useState } from 'react';
 import { useDispatch } from 'react-redux';
+import { authService } from '../firebase';
 import {
   deleteComment,
   getComment,
@@ -7,7 +8,7 @@ import {
 } from '../redux/modules/commentModule';
 import Button from './Button';
 
-function Comment({ postId, commentId, body, creator }) {
+function Comment({ postId, commentId, body, userUID, displayName }) {
   const [toggle, setToggle] = useState(false);
   const [newBody, setNewBody] = useState('');
   const dispatch = useDispatch();
@@ -16,6 +17,10 @@ function Comment({ postId, commentId, body, creator }) {
     dispatch(getComment(postId));
   }, [toggle]);
   const onClickToggle = () => {
+    if (userUID !== authService?.currentUser?.uid) {
+      alert('로그인을 해주세요.');
+      return;
+    }
     setToggle(!toggle);
   };
 
@@ -30,6 +35,10 @@ function Comment({ postId, commentId, body, creator }) {
   };
 
   const deleteClickComment = () => {
+    if (userUID !== authService?.currentUser?.uid) {
+      alert('로그인을 해주세요.');
+      return;
+    }
     if (window.confirm('정말 삭제하시겠습니까?')) {
       dispatch(deleteComment({ postId, commentId }));
       alert('삭제되었습니다.');
@@ -37,7 +46,7 @@ function Comment({ postId, commentId, body, creator }) {
   };
   return (
     <div>
-      <div>{creator}</div>
+      <div>{displayName ?? '익명사용자'}</div>
       {!toggle ? (
         <div>{body}</div>
       ) : (
@@ -52,8 +61,12 @@ function Comment({ postId, commentId, body, creator }) {
           <Button>수정완료</Button>
         </form>
       )}
-      <Button onClick={onClickToggle}>{toggle ? '취소' : '수정'}</Button>
-      <Button onClick={deleteClickComment}>삭제</Button>
+      {userUID === authService?.currentUser?.uid ? (
+        <div>
+          <Button onClick={onClickToggle}>{toggle ? '취소' : '수정'}</Button>
+          <Button onClick={deleteClickComment}>삭제</Button>
+        </div>
+      ) : null}
       <hr />
     </div>
   );
